--- src/pages/Privacy.jsx
@@ -1,476 +1,0 @@
-
-import React from 'react'
-import "./privacy.css"
-
-
-
-export default function Privacy() {
-  return (
-		<article>
-			<h1 className='privacy-text__h1'>Privacy Policy for Finosell</h1>
-			<h3 className='privacy-text__h3'>
-				<strong>Privacy Policy</strong>
-			</h3>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong>Last updated: August 03, 2022</strong>
-			</h5>
-			<p>
-				This Privacy Policy describes Our policies and procedures on the
-				collection, use, and disclosure of Your information when You use the
-				Service and tells You about Your privacy rights and how the law protects
-				you.
-			</p>
-			<p>
-				We use Your Personal data to provide and improve the Service. By using
-				the Service, You agree to the collection and use of information in
-				accordance with this Privacy Policy.
-			</p>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong>Interpretation and Definitions</strong>
-			</h5>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong>Interpretation</strong>
-			</h5>
-			<p>
-				The words in which the initial letter is capitalized have meanings
-				defined under the following conditions. The following definitions shall
-				have the same meaning regardless of whether they appear in singular or
-				in the plural.
-			</p>
-			<h5 className='privacy-text__h5'>
-				<strong>Definitions </strong>
-			</h5>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong> For this Privacy Policy:</strong>
-			</h5>
-			<br />
-			<p>
-				{' '}
-				<strong> Account</strong> means a unique account created for You to
-				access our Service or parts of our Service. Affiliate means an entity
-				controlled by or under common control with a party, where "control"
-				means ownership of 50% or more of the shares, equity interest, or other
-				securities entitled to vote for the election of directors or other
-				managing authority.
-			</p>
-			<br />{' '}
-			<p>
-				<strong> Application</strong> means the software program provided by the
-				Company downloaded by You on any electronic device named Finosell.
-			</p>
-			<br />
-			<p>
-				<strong> Company</strong> (referred to as either "the Company," "We",
-				"Us," or "Our" in this Agreement) refers to Finosell.
-			</p>
-			<br />
-			<p>
-				<strong>Cookies</strong> are small files placed on Your computer, mobile
-				device, or any other device by a website, containing the details of Your
-				browsing history on that website among its many uses.
-			</p>
-			<br />
-			<p>
-				<strong>Country</strong> refers to Nigeria.
-			</p>
-			<br />
-			<p>
-				<strong>Device</strong> means any device that can access the Service,
-				such as a computer, a cellphone, or a digital tablet.
-			</p>
-			<br />
-			<p>
-				<strong> Personal Data</strong> is any information that relates to an
-				identified or identifiable individual.
-				<br />
-				<strong>Service</strong> refers to the Application or the Website, or
-				both.
-			</p>
-			<br />
-			<p>
-				<strong>Service Provider</strong> means any natural or legal person who
-				processes the data on behalf of the Company. It refers to third-party
-				companies or individuals employed by the Company to facilitate the
-				Service, to provide the Service on behalf of the Company, to perform
-				services related to the Service, or to assist the Company in analyzing
-				how the Service is used.
-			</p>
-			<br />
-			<p>
-				<strong>Usage Data</strong> refers to data collected automatically,
-				either generated by the Service or from the Service infrastructure
-				itself (for example, the duration of a page visit).
-				<br />
-				<strong> Website</strong> refers to Finosell, accessible from
-				www.finosell.com <br />
-				<strong>You</strong> mean the individual accessing or using the Service,
-				the company, or other legal entity on behalf of which such individual is
-				accessing or using the Service, as applicable.
-			</p>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong>Collecting and Using Your Personal Data</strong>
-			</h5>
-			<h5 className='privacy-text__h5'>
-				<strong>Types of Data Collected</strong>
-			</h5>
-			<h5 className='privacy-text__h5'>
-				<strong>Personal Data</strong>
-			</h5>{' '}
-			<p>
-				While using Our Service, We may ask You to provide Us with certain
-				personally identifiable information that can be used to contact or
-				identify You. Personally, identifiable information may include, but is
-				not limited to:
-			</p>
-			<ul className='privacy-text__ul'>
-				<li>Email address</li>
-				<li>First name and last name</li>
-				<li>Phone number</li>
-				<li>Address, State, Province, ZIP/Postal code, City</li>
-				<li>Usage Data</li>
-			</ul>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong> Usage Data</strong>
-			</h5>
-			<p>
-				Usage Data is collected automatically when using the Service. Usage Data
-				may include information such as Your Device's Internet Protocol address
-				(e.g., IP address), browser type, browser version, the pages of our
-				Service that You visit, the time and date of Your visit, the time spent
-				on those pages, unique device identifiers and other diagnostic data.
-			</p>
-			<p>
-				When You access the Service by or through a mobile device, We may
-				collect certain information automatically, including, but not limited
-				to, the type of mobile device You use, Your mobile device's unique ID,
-				the IP address of Your mobile device, Your mobile operating system, the
-				type of mobile Internet browser You use, unique device identifiers and
-				other diagnostic data. We may also collect information that Your browser
-				sends whenever You visit our Service or when You access the Service by
-				or through a mobile device.
-			</p>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong>Information Collected while Using the Application </strong>
-			</h5>
-			<p>
-				While using Our Application, to provide features of Our Application,
-				<br /> We may collect, with Your prior permission:
-			</p>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong>Information regarding your location </strong>
-			</h5>
-			<p>
-				We use this information to provide features of Our Service and to
-				improve and customize Our Service. The information may be uploaded to
-				the Company's servers and/or a Service Provider's server or stored on
-				Your device. You can enable or disable access to this information
-				anytime, through Your Device settings. Tracking Technologies and Cookies
-				We use Cookies and similar tracking technologies to track the activity
-				on Our Service and store certain information. Tracking technologies are
-				beacons, tags, and scripts to collect and track information and improve
-				and analyze Our Service.
-			</p>
-			<br />
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong>The technologies We use may include: </strong>
-			</h5>
-			<h5 className='privacy-text__h5'>
-				<strong>Cookies or Browser Cookies.</strong>
-			</h5>
-			<p>
-				A cookie is a small file placed on Your Device. You can instruct Your
-				browser to refuse all Cookies or to indicate when a cookie is being
-				sent. However, if You do not accept Cookies, You may not be able to use
-				some parts of our Service. Unless you have adjusted Your browser setting
-				so it will refuse Cookies, our Service may use Cookies.
-			</p>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong> Flash Cookies.</strong>
-			</h5>
-			<p>
-				Certain features of our Service may use local stored objects (or Flash
-				Cookies) to collect and store information about Your preferences or
-				activity on our Service. Flash Cookies are not managed by the same
-				browser settings as those used for Browser Cookies.
-			</p>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong> Web Beacons. </strong>
-			</h5>
-			<p>
-				Certain sections of our Service and our emails may contain small
-				electronic files known as web beacons(also referred to as clear gifs,
-				pixel tags, and single-pixel gifs)that permit the Company, for example,
-				to count users who have visited those pages or opened an email and for
-				other related website statistics (for example, recording the popularity
-				of a certain section and verifying system and server integrity).
-			</p>
-			<br />
-			<p>
-				<strong>Cookies can be "Persistent" or "Session" Cookies.</strong>{' '}
-				Persistent Cookies remain on Your personal computer or mobile device
-				when You go offline, while Session Cookies are deleted as soon as You
-				close Your web browser. Learn more about cookies on the Free Privacy
-				Policy website article.{' '}
-			</p>
-			<p>
-				We use both Session, and Persistent Cookies for the purposes set out
-				below:
-			</p>
-			<p>Necessary / Essential Cookies</p>
-			<p>Type: Session Cookies</p>
-			<p> Administered by: Us</p>
-			<br />
-			<p>
-				<strong>Purpose:</strong> These Cookies are essential to provide You
-				with services available through the Website and to enable You to use
-				some of its features. They help to authenticate users and prevent
-				fraudulent use of user accounts. Without these Cookies, the services
-				that You have asked for cannot be provided, and We only use these
-				Cookies to provide You with those services.
-			</p>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong>Use of Your Personal Data </strong>
-			</h5>
-			<p>
-				{' '}
-				The Company may use Personal Data for the following purposes: To provide
-				and maintain our Service, including monitoring the usage of our Service.
-				To manage Your Account: to manage Your registration as a user of the
-				Service. The Personal Data You provide can give You access to different
-				functionalities of the Service that are available to You as a registered
-				user.
-			</p>
-			<br />
-			<p>
-				For the performance of a contract: the development, compliance, and
-				undertaking of the purchase contract for the products, items, or
-				services You have purchased or of any other contract with Us through the
-				Service.
-			</p>
-			<br />
-			<p>
-				To contact You: To contact You by email, telephone calls, SMS, or other
-				equivalent forms of electronic communication, such as a mobile
-				application's push notifications regarding updates or informative
-				communications related to the functionalities, products or contracted
-				services, including the security updates, when necessary or reasonable
-				for their implementation.
-			</p>
-			<br />
-			<p>
-				To provide You with news, special offers, and general information about
-				other goods, services, and events which we offer that are similar to
-				those that you have already purchased or enquired about unless You have
-				opted not to receive such information.
-			</p>
-			<br />
-			<p>To manage Your requests: To attend and manage Your requests to Us. </p>
-			<br />
-			<p>
-				For business transfers: We may use Your information to evaluate or
-				conduct a merger, divestiture, restructuring, reorganization,
-				dissolution, or another sale or transfer of some or all of Our assets,
-				whether as a going concern or as part of bankruptcy, liquidation, or
-				similar proceeding, in which Personal Data held by Us about our Service
-				users is among the assets transferred.
-			</p>{' '}
-			<p>
-				For other purposes: We may use Your information for other purposes, such
-				as data analysis, identifying usage trends, determining the
-				effectiveness of our promotional campaigns, and to evaluate and
-				improving our Service, products, services, marketing, and your
-				experience.
-			</p>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong>
-					{' '}
-					We may share Your personal information in the following situations:{' '}
-				</strong>
-			</h5>
-			<p>
-				With Service Providers: We may share Your personal information with
-				Service Providers to monitor and analyze the use of our Service to
-				contact You.
-			</p>
-			<p>
-				For business transfers: We may share or transfer Your personal
-				information in connection with, or during negotiations of, any merger,
-				sale of Company assets, financing, or acquisition of all or a portion of
-				Our business to another company.
-			</p>
-			<p>
-				With Affiliates: We may share Your information with Our affiliates,
-				which will require those affiliates to honor this Privacy Policy.
-				Affiliates include Our parent company and any other subsidiaries, joint
-				venture partners or companies that We control or are under common
-				control with Us.
-			</p>
-			<p>
-				With business partners: We may share Your information with Our business
-				partners to offer You certain products, services, or promotions.
-			</p>
-			<p>
-				With Your consent: We may disclose Your personal information for any
-				other purpose with Your consent.
-			</p>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong> Retention of Your Personal Data </strong>
-			</h5>
-			<p>
-				The Company will retain Your Personal Data only for as long as is
-				necessary for the purposes set out in this Privacy Policy. We will
-				retain and use Your Data to the extent necessary to comply with our
-				legal obligations (for example, if we are required to retain your data
-				to comply with applicable laws), resolve disputes, and enforce our legal
-				agreements and policies.
-			</p>
-			<p>
-				The Company will also retain Usage Data for internal analysis purposes.
-				Usage Data is generally retained for a shorter time, except when this
-				data is used to strengthen the security or improve the functionality of
-				Our Service. We are legally obligated to retain this data for longer
-				periods.
-			</p>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong>Transfer of Your Personal Data </strong>
-			</h5>
-			<p>
-				Your information, including Personal Data, is processed at the Company's
-				operating offices and in any other places where the parties involved in
-				the processing are located. It means that this information may be
-				transferred to — and maintained on — computers located outside of Your
-				state, province, country, or other governmental jurisdiction where the
-				data protection laws may differ from those of Your jurisdiction.
-			</p>
-			<p>
-				Your consent to this Privacy Policy followed by Your submission of such
-				information represents Your agreement to that transfer.
-			</p>
-			<p>
-				The Company will take all steps reasonably necessary to ensure that Your
-				data is treated securely and in accordance with this Privacy Policy, and
-				no transfer of Your Personal Data will take place to an organization or
-				a country unless there are adequate controls in place, including the
-				security of Your data and other personal information.
-			</p>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong> Disclosure of Your Personal Data</strong>
-			</h5>
-			<h5 className='privacy-text__h5'>
-				<strong> Business Transactions </strong>
-			</h5>
-			If the Company is involved in a merger, acquisition or asset sale, Your
-			Personal Data may be transferred. We will provide notice before Your
-			Personal Data is transferred and subject to a different Privacy Policy.
-			<h5 className='privacy-text__h5'>
-				<strong>Law enforcement</strong>
-			</h5>
-			<p>
-				{' '}
-				Under certain circumstances, the Company may be required to disclose
-				Your Data if required by law or in response to valid requests by public
-				authorities (e.g., a court or a government agency).
-			</p>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong> Other Legal Requirements</strong>
-			</h5>
-			<p>
-				{' '}
-				The Company may disclose Your Data in the good faith belief that such
-				action is necessary to:
-			</p>
-			<ul className='privacy-text__ul'>
-				<li>Comply with a legal obligation</li>
-				<li> Protect and defend the rights or property of the Company</li>
-				<li>
-					{' '}
-					Prevent or investigate possible wrongdoing in connection with the
-					Service
-				</li>
-				<li>
-					Protect the personal safety of Users of the Service or the public
-				</li>
-				<li>Protect against legal liability</li>
-			</ul>
-			<h5 className='privacy-text__h5'>
-				<strong>Security of Your Personal Data</strong>
-			</h5>
-			<p>
-				The security of Your Personal Data is important to Us. We strive to use
-				commercially acceptable means to protect Your Personal Data.
-			</p>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong>Children's Privacy</strong>
-			</h5>
-			<p>
-				Our Service does not address anyone under the age of 13. We do not
-				knowingly collect personally identifiable information from anyone under
-				13. If You are a parent or guardian and know that Your child has
-				provided Us with Personal Data, please contact Us. Suppose We become
-				aware that We have collected Personal Data from anyone under the age of
-				13 without verification of parental consent. In that case, We take steps
-				to remove that information from Our servers.
-			</p>
-			<p>
-				If We need to rely on consent as a legal basis for processing Your
-				information and Your country requires consent from a parent, We may
-				require Your parent's permission before We collect and use that
-				information.
-			</p>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong>Links to Other Websites </strong>
-			</h5>
-			<p>
-				Our Service may contain links to other websites that we do not operate.
-				If You click on a third-party link, You will be directed to that third
-				party's site. We strongly advise You to review the Privacy Policy of
-				every site You visit. We have no control over and assume no
-				responsibility for the content, privacy policies, or practices of any
-				third-party sites or services.
-			</p>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong> Changes to this Privacy Policy </strong>
-			</h5>
-			<p>
-				We may update Our Privacy Policy from time to time. We will notify You
-				of any changes by posting the new Privacy Policy on this page. We will
-				inform you via email or a prominent notice on Our Service before the
-				change becomes effective and update the "Last updated" date at the top
-				of this Privacy Policy.
-			</p>
-			<br />
-			<h5 className='privacy-text__h5'>
-				<strong>Contact Us</strong>
-			</h5>
-			<p>
-				If you have any questions about this Privacy Policy, You can contact us:{' '}
-				<br />
-				By email: support@finosell.com
-			</p>
-		</article>
-	);
-}
-
-
-
-
