--- src/pages/escrow.css
@@ -1,13 +1,0 @@
-article {
-    margin: 80px;
-}
-
-
-
-.escrow-text__h1 {
-    font-family: DM Sans;
-        font-size: 35px;
-        font-weight: 900;
-        text-align: center;
-        margin-bottom: 50px;
-}
